T #93 moved job manager to services
